requirements.txt
fastapi
uvicorn
requests
pymupdf
faiss-cpu
numpy
python-multipart
pillow
sentence-transformers
google-generativeai
openai
